requirements.txt
asgiref==3.8.1
channels==4.2.0
channels_redis==4.2.1
daphne==4.1.2
Django==5.1.5
django-cors-headers==4.6.0
django-environ==0.12.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
Faker==35.1.0
psycopg2-binary==2.9.10
redis==5.2.1
